api/requirements.txt
alembic==1.6.5
argon2-cffi==19.1.0
asn1crypto==0.24.0
celery[redis]<=5.1
click<8
cryptography==3.4.7
email-validator==1.1.3
Flask==1.1.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.4.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
Flask-Marshmallow==0.14.0
flower
gunicorn==19.9.0
importlib-metadata
marshmallow-sqlalchemy==0.26.1
numpy
numpy-stl==2.16.0
passlib==1.7.4
Pillow==8.3.1
PyMySQL==0.9.3
pytest==6.2.4
pytest-cov==2.12.1
python-dotenv==0.18.0
python-magic==0.4.24
python-openctm==1.0.11
requests==2.26.0
SQLAlchemy==1.4.21
trimesh==3.9.21
urllib3==1.26.6
wcwidth==0.2.5
Werkzeug==0.15.6
wheel
zipp==0.5.1
